Refactoring and bug fix
Worked around JIT bug, which was breaking nested line diagrams after some
time of usage

Speed up building of nested line diagrams

Extracted DiagramHistory in separate class (needs some more refactoring),
added visitor pattern for visiting all zoomed concepts

Added first test case (very insignificant one) for fixing bug for
mapping of mouse position to diagram one.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@174 267469b6-68fb-4578-8a4b-5844dc4013a7

+++ source/net/sourceforge/toscanaj/model/diagram/NestedDiagramNode.java
@@ -28,8 +28,8 @@ public class NestedDiagramNode extends DiagramNode {
      * If the dropAttributeLabels flag is set, the inner nodes will have no
      * attribute labels attached.
      */
-    public NestedDiagramNode(DiagramNode outerNode, Diagram2D innerDiagram, double scale,
-                                         boolean dropAttributeLabels ) {
+    public NestedDiagramNode(DiagramNode outerNode, Diagram2D innerDiagram, float scale,
+                                        boolean dropAttributeLabels ) {
         super( new Point2D.Double(outerNode.getX()*scale, outerNode.getY()*scale),
                outerNode.getConcept(),
                new LabelInfo(outerNode.getAttributeLabelInfo()), null);
